feat(mdlib): add BaseDirectiveBlock; add unittests
directives are probably to me the best designed feature of the rst language. they allow to include consistent custom feature to the language while being very flexible.

the supported syntax is probably not a 1:1 match with the rst spec but is close enough that I won't probably need to edit the directives syntax from rst document to have it work in markdown.